Changes to the graphics mainly.
Added a background to the main tab.
Added a .ico to make it a bit more personalised.
Changed the size of the progress bar (moved also) *looking into tweaking
it more
Moved check for updates button
Changed 'play' button, moved it. However encountered a problem here
where I couldn't keep the transparency of the button image and have it
fit into the background and keep the functionality of the button itself,
it still sends a request to launch the client, but it acts more like a
hyperlink than a button. Because of the Flatstyle option, standard ads a
background to the button and ignores transparency in the image, windows
is lame (for lack of a better term) and popup ads a border but keeps the
transparency, but the border alone ruins the aesthetics. This I will
need a bit of a hand with nutting out how to keep the style of the
button (should we decide to keep it that way(if you like it)) yet make
it act more like a button so when pressed it actually feels as though
you pressed something. So i'm looking, in essence, to alter the feedback
given from a mouseclick or onclick.